:truck: Move Azure authentication from mixin to external
