+++ data_safe_haven/administration/users/active_directory_users.py
@@ -5,7 +5,7 @@ from typing import Any, Optional, Sequence
 
 # Local imports
 from data_safe_haven.config import Config
-from data_safe_haven.external.api import AzureApi
+from data_safe_haven.external import AzureApi
 from data_safe_haven.functions import b64encode
 from data_safe_haven.pulumi import PulumiStack
 from data_safe_haven.utility import FileReader, Logger
