+++ data_safe_haven/administration/users/azure_ad_users.py
@@ -3,7 +3,7 @@
 from typing import Any, Sequence
 
 # Local imports
-from data_safe_haven.external.api import GraphApi
+from data_safe_haven.external import GraphApi
 from data_safe_haven.functions import password
 from data_safe_haven.utility import Logger
 from .research_user import ResearchUser
