+++ data_safe_haven/backend/backend.py
@@ -5,7 +5,7 @@ from typing import Any, Optional
 # Local imports
 from data_safe_haven.config import Config, DotFileSettings
 from data_safe_haven.exceptions import DataSafeHavenAzureException
-from data_safe_haven.external.api import AzureApi
+from data_safe_haven.external import AzureApi
 
 
 class Backend:
