+++ data_safe_haven/commands/deploy_shm_command.py
@@ -16,7 +16,7 @@ from data_safe_haven.exceptions import (
     DataSafeHavenException,
     DataSafeHavenInputException,
 )
-from data_safe_haven.external.api import GraphApi
+from data_safe_haven.external import GraphApi
 from data_safe_haven.functions import (
     password,
     validate_aad_guid,
