+++ data_safe_haven/external/__init__.py
@@ -1,0 +1,19 @@
+from .api.azure_api import AzureApi
+from .api.azure_cli import AzureCli
+from .api.graph_api import GraphApi
+from .interface.azure_authenticator import AzureAuthenticator
+from .interface.azure_container_instance import AzureContainerInstance
+from .interface.azure_fileshare import AzureFileShare
+from .interface.azure_ipv4_range import AzureIPv4Range
+from .interface.azure_postgresql_database import AzurePostgreSQLDatabase
+
+__all__ = [
+    "AzureApi",
+    "AzureAuthenticator",
+    "AzureCli",
+    "AzureContainerInstance",
+    "AzureFileShare",
+    "AzureIPv4Range",
+    "AzurePostgreSQLDatabase",
+    "GraphApi",
+]
