--- data_safe_haven/external/api/__init__.py
@@ -1,9 +1,0 @@
-from .azure_api import AzureApi
-from .azure_cli import AzureCli
-from .graph_api import GraphApi
-
-__all__ = [
-    "AzureApi",
-    "AzureCli",
-    "GraphApi",
-]
