--- data_safe_haven/external/interface/__init__.py
@@ -1,11 +1,0 @@
-from .azure_container_instance import AzureContainerInstance
-from .azure_fileshare import AzureFileShare
-from .azure_ipv4_range import AzureIPv4Range
-from .azure_postgresql_database import AzurePostgreSQLDatabase
-
-__all__ = [
-    "AzureContainerInstance",
-    "AzureFileShare",
-    "AzureIPv4Range",
-    "AzurePostgreSQLDatabase",
-]
