--- data_safe_haven/mixins/__init__.py
@@ -1,5 +1,0 @@
-from .azure_mixin import AzureMixin
-
-__all__ = [
-    "AzureMixin",
-]
