+++ data_safe_haven/provisioning/shm_provisioning_manager.py
@@ -1,6 +1,6 @@
 """Provisioning manager for a deployed SHM."""
 # Local imports
-from data_safe_haven.external.api import AzureApi
+from data_safe_haven.external import AzureApi
 from data_safe_haven.pulumi import PulumiStack
