+++ data_safe_haven/pulumi/components/shm_networking.py
@@ -7,7 +7,7 @@ from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import network, resources
 
 # Local imports
-from data_safe_haven.external.interface import AzureIPv4Range
+from data_safe_haven.external import AzureIPv4Range
 from data_safe_haven.functions import ordered_private_dns_zones
 from ..common.enums import NetworkingPriorities
