+++ data_safe_haven/pulumi/components/sre_data.py
@@ -14,7 +14,7 @@ from pulumi_azure_native import (
 )
 
 # Local imports
-from data_safe_haven.external.interface import AzureIPv4Range
+from data_safe_haven.external import AzureIPv4Range
 from data_safe_haven.functions import (
     alphanumeric,
     ordered_private_dns_zones,
