+++ data_safe_haven/pulumi/components/sre_remote_desktop.py
@@ -14,7 +14,7 @@ from pulumi_azure_native import (
 )
 
 # Local imports
-from data_safe_haven.external.interface import AzureIPv4Range
+from data_safe_haven.external import AzureIPv4Range
 from data_safe_haven.pulumi.common.transformations import (
     get_id_from_subnet,
     get_ip_address_from_container_group,
