+++ data_safe_haven/pulumi/dynamic/blob_container_acl.py
@@ -8,7 +8,7 @@ from pulumi.dynamic import CreateResult, DiffResult, Resource
 
 # Local imports
 from data_safe_haven.exceptions import DataSafeHavenPulumiException
-from data_safe_haven.external.api import AzureApi
+from data_safe_haven.external import AzureApi
 from .dsh_resource_provider import DshResourceProvider
