+++ data_safe_haven/pulumi/dynamic/remote_powershell.py
@@ -7,7 +7,7 @@ from pulumi import Input, Output, ResourceOptions
 from pulumi.dynamic import CreateResult, DiffResult, Resource
 
 # Local imports
-from data_safe_haven.external.api import AzureApi
+from data_safe_haven.external import AzureApi
 from .dsh_resource_provider import DshResourceProvider
