+++ data_safe_haven/pulumi/pulumi_stack.py
@@ -14,7 +14,7 @@ from pulumi import automation
 # Local imports
 from data_safe_haven.config import Config
 from data_safe_haven.exceptions import DataSafeHavenPulumiException
-from data_safe_haven.external.api import AzureCli
+from data_safe_haven.external import AzureCli
 from data_safe_haven.utility import Logger
 from .declarative_shm import DeclarativeSHM
 from .declarative_sre import DeclarativeSRE
